requirements.txt
RPi.GPIO
pymodbus
